requirements.txt
# PDF Generation
reportlab==4.0.9
pdfplumber==0.11.0
PyPDF2==3.0.1

# LLM Integration
openai>=1.40.0

# Configuration
python-dotenv==1.0.0

# Data Export
PyYAML==6.0.1

# CLI
colorama==0.4.6
tqdm==4.66.1

# Web UI
gradio>=4.0.0
